Move SharedSelectionViewModel to ui.resources
- Relocated `SharedSelectionViewModel` from `ui.shared` to `ui.resources`
- Updated the package declaration and imports accordingly
